CB-12407: Increase FreeIPA certificate serial number max value
The FreeIPA certificate serial number's is an ASN.1 integer. ANS.1
intergers are unbounded. The java code was increased from an int to a
long. While this doesn't fix the problem in all cases, in practise it
should be sufficient. When FreeIPA is installed, it will pick a range
and use consecutive serial number for that instance of FreeIPA.

This was tested using a local deployment of cloudbreak that had an out
of range number in the serial number.

+++ freeipa-client/src/main/java/com/sequenceiq/freeipa/client/FreeIpaClient.java
@@ -267,7 +267,7 @@ public class FreeIpaClient {
         return (Set<Cert>) invoke("cert_find", List.of(), Map.of(), type).getResult();
     }
 
-    public void revokeCert(int serialNumber) throws FreeIpaClientException {
+    public void revokeCert(long serialNumber) throws FreeIpaClientException {
         List<Object> flags = List.of(String.valueOf(serialNumber));
         Map<String, Object> params = Map.of("revocation_reason", CESSATION_OF_OPERATION);
         invoke("cert_revoke", flags, params, Object.class);
